+++ scripts/action-handler.js
@@ -1,5 +1,5 @@
 // System Module Imports
-import { ACTION_TYPE, ITEM_TYPE } from './constants.js'
+import { ACTION_TYPE, GROUP } from './constants.js'
 import { Utils } from './utils.js'
 
 export let ActionHandler = null
@@ -17,27 +17,27 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
          */a
         async buildSystemActions(groupIds) {
             // Set actor and token variables
-            this.actors = (!this.actor) ? this.#getActors() : [this.actor]
-            this.tokens = (!this.token) ? this.#getTokens() : [this.token]
+            this.actors = (!this.actor) ? this.#getValidActors() : [this.actor]
+            this.tokens = (!this.token) ? this.#getValidTokens() : [this.token]
 
-            this.actorType = this.actor?.type
-
+            // Set items variable
+            if (this.actor) {
+                this.items = coreModule.api.Utils.sortItemsByName(this.actor.items);
+            }
 
             // Settings
             this.displayUnequipped = Utils.getSetting('displayUnequipped')
 
-            // Set items variable
-            if (this.actor) {
-                let items = this.actor.items
-                items = coreModule.api.Utils.sortItemsByName(items)
-                this.items = items
-            }
-
-            if (this.actorType === 'character' || this.actorType === 'minion' || this.actorType === 'rival' || this.actorType === 'nemesis' || this.actorType === 'vehicle') {
+            if (this.actor?.type === 'character' || this.actor?.type === 'minion' || this.actor?.type === 'rival' || this.actor?.type === 'nemesis') {
                 this.inventorygroupIds = [
                     'weapons'
                 ]
                 await this.#buildCharacterActions()
+            } else if (this.actor?.type === "vehicle") {
+                this.inventorygroupIds = [
+                    'weapons'
+                ]
+                await this.#buildVehicleActions();
             } else if (!this.actor) {
                 this.#buildMultipleTokenActions()
             }
@@ -46,17 +46,25 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
         /**
          * Build character actions
          * @private
+         * @returns {object}
          */
         async #buildCharacterActions() {
             await Promise.all([
                 this.#buildInventory(),
             ])
-            if (this.actorType !== 'vehicle') {
-                this.#buildSkills()
-            } else {
-                this.#buildCrewSkills()
-            }
+            this.#buildSkills()
+        }
 
+        /**
+         * Build vehicle actions
+         * @private
+         * @returns {object}
+         */
+        async #buildVehicleActions() {
+            await Promise.all([
+                this.#buildInventory(),
+            ])
+            this.#buildCrewSkills()
         }
 
         /**
@@ -65,6 +73,7 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
          * @returns {object}
          */
         #buildMultipleTokenActions() {
+            this.#buildSkills();
         }
 
         /**
@@ -74,47 +83,41 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
         async #buildInventory() {
             if (this.items.size === 0) return
 
-            const actionTypeId = 'item'
-            const inventoryMap = new Map()
+            let actionType = '';
+            const inventoryMap = new Map([
+                ["weapons", new Map()]
+            ]);
 
             for (const [itemId, itemData] of this.items) {
-                const type = itemData.type
-                const equipped = itemData.system?.equippable?.equipped
 
-                if (equipped || this.displayUnequipped) {
-                    const typeMap = inventoryMap.get(type) ?? new Map()
-                    typeMap.set(itemId, itemData)
-                    inventoryMap.set(type, typeMap)
+                if (this.#isEquippedItem(itemData)) {
+                    switch (itemData.type) {
+                        case "shipweapon":
+                            actionType = "shipweapon"
+                            inventoryMap.get("weapons").set(itemId, itemData);
+                            break;
+                        case "weapon":
+                            actionType = "weapon"
+                            inventoryMap.get("weapons").set(itemId, itemData);
+                            break;
+                    }
                 }
             }
 
-            for (const [type, typeMap] of inventoryMap) {
-                const groupId = ITEM_TYPE[type]?.groupId
+            for (const groupId of this.inventorygroupIds) {
+                const actionData = inventoryMap.get(groupId);
+                if (!actionData || actionData.size === 0) continue;
 
-                if (!groupId) continue
+                // Create group data
+                const groupData = {
+                    id: groupId,
+                    name: game.i18n.localize(GROUP[groupId].name)
+                };
 
-                const groupData = { id: groupId, type: 'system' }
+                const data = { groupData, actionData, actionType };
+                // Build actions and activations
+                await this.buildActions(data);
 
-                // Get actions
-                const actions = [...typeMap].map(([itemId, itemData]) => {
-                    const id = itemId
-                    const name = itemData.name
-                    const actionTypeName = coreModule.api.Utils.i18n(ACTION_TYPE[actionTypeId])
-                    const listName = `${actionTypeName ? `${actionTypeName}: ` : ''}${name}`
-                    const encodedValue = [actionTypeId, id].join(this.delimiter)
-                    const img = coreModule.api.Utils.getImage(itemData.img)
-
-                    return {
-                        id,
-                        name,
-                        img,
-                        listName,
-                        encodedValue
-                    }
-                })
-
-                // TAH Core method to add actions to the action list
-                this.addActions(actions, groupData)
             }
         }
 
@@ -122,46 +125,40 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
          * Build skills
          * @private
          */
-        #buildSkills() {
-            //if (this.actorType !== 'character') return
+        async #buildSkills() {
 
-            const skills = this.actor.system.skills
+            const skills = this.actor?.system.skills
+            if (skills.length === 0) return;
+
             const actionType = 'skill'
             const categoriesSkillsList = this.actor.system.skilltypes
-            for (const [actionId, category] of Object.entries(categoriesSkillsList)) {
-                try {
-                // Create group data  SWFFG.SkillsCombat
-                const groupData = {
-                    id: category.type,
-                    name: category.label,
-                    type: 'system'
-                }
-                const categorizedSkills = Object.entries(skills).filter(skill => skill[1].type === category.type)
-                const actions = Object.entries(categorizedSkills).map((skill) => {
-                    const id = skill[1][1].value
-                    const encodedValue = [actionType, id].join(this.delimiter)
-                    const name = skill[1][1].label
-                    const actionTypeName = `${coreModule.api.Utils.i18n('SWFFG.Skills')}: ` ?? ''
-                    const listName = `${actionTypeName}${name}`
-                    const tooltip = "ddd"
-        
-                    return {
-                      id,
-                      name,
-                      encodedValue,
-                      listName,
-                      tooltip
-                    }
-                  })
-                // Add actions to HUD
-                this.addActions(actions, groupData)
-                } catch (error) {
-                    coreModule.api.Logger.error(actionId)
-                    return null
+            const skillMap = new Map([
+                ["General", new Map()],
+                ["Social", new Map()],
+                ["Knowledge", new Map()],
+                ["Combat", new Map()]
+            ]);
+
+            for (const [skillId, skillData] of Object.entries(skills)) {
+                if (skillMap.get(skillData.type)) {
+                    skillMap.get(skillData.type).set(skillId, skillData);
                 }
 
             }
+            for (const [groupId, categoryData] of Object.entries(categoriesSkillsList)) {
+                const actionData = skillMap.get(categoryData.type);
+                if (!actionData || actionData.size === 0) continue;
 
+                // Create group data
+                const groupData = {
+                    id: categoryData.type,
+                    name: game.i18n.localize(GROUP[categoryData.type].name)
+                };
+
+                const data = { groupData, actionData, actionType };
+                // Build actions and activations
+                this.buildActions(data);
+            }
         }
 
         /**
@@ -169,59 +166,37 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
          * @private
          */
         async #buildCrewSkills() {
-            //if (this.actorType !== 'character') return
             const ship = this.actor;
-            const crew = await ship.getFlag("starwarsffg", "crew");
+            const crew = await this.actor.getFlag("starwarsffg", "crew");
             if (!crew || crew.length === 0) {
                 CONFIG.logger.warn("Could not find crew for vehicle or could not find relevant skill; presenting default roller");
                 return;
             }
-            const actionType = 'skill'
+            const actionType = 'crewSkill'
             // Create group data  SWFFG.SkillsCombat
             const groupData = {
                 id: "crewskills",
-                name: "category.label",
+                name: game.i18n.localize(GROUP["crewskills"].name),
                 type: 'system'
             }
             for (const [actionId, values] of Object.entries(crew)) {
                 try {
                     // Create actions list
-                    let actions = new Array()
-                    let skillRole = "";
-                    let skill = "";
-                    let use_handling = false;
-                    if (values.role === 'Pilot') {
-                        use_handling = true;
-                        if (ship?.system?.spaceShip) {
-                            skill = "Piloting: Space";
-                        } else {
-                            skill = "Piloting: Planetary";
-                        }
-                    } else {
-                        skillRole = game.settings.get("starwarsffg", "arrayCrewRoles").filter(role => role.role_name === values.role);
-                        skill = skillRole[0].role_skill
-                        use_handling = skillRole[0].use_handling
-                    }
-
+                    const actions = new Array()
                     const name = values.role
-                    const id = [skill, actionId].join('-');
-                    const crewActorId = values.actor_id
-                    const crewActor = game.actors.get(crewActorId);
-                    const encodedValue = [actionType, skill, crewActorId, use_handling].join(this.delimiter)
-                    const actionTypeName = `${coreModule.api.Utils.i18n('SWFFG.Skills')}: ` ?? ''
-                    const listName = `${actionTypeName}${name}`
-                    const img = coreModule.api.Utils.getImage(crewActor.img)
+                    const id = [values.actor_id, values.role].join(this.delimiter);
+                    const listName = this.#getListName(actionType, name)
+                    const img = coreModule.api.Utils.getImage(game.actors.get(values.actor_id).img)
                     const tooltip = values.actor_name
-
-
                     actions.push({
                         id: id,
                         name: name,
-                        encodedValue: encodedValue,
                         listName: listName,
                         img: img,
                         tooltip: tooltip,
+                        system: { actionType, actionId: id }
                     })
+
                     // Add actions to HUD
                     this.addActions(actions, groupData)
                 } catch (error) {
@@ -233,35 +208,182 @@ Hooks.once('tokenActionHudCoreApiReady', async (coreModule) => {
         }
 
         /**
-        * Get actors
-        * @private
-        * @returns {object}
-        */
-        #getActors() {
-            const allowedTypes = ['character', 'npc']
-            const tokens = coreModule.api.Utils.getControlledTokens()
-            const actors = tokens?.filter(token => token.actor).map((token) => token.actor)
-            if (actors.every((actor) => allowedTypes.includes(actor.type))) {
-                return actors
-            } else {
-                return []
-            }
-        }
-
-        /**
-         * Get tokens
+         * Get valid actors
          * @private
          * @returns {object}
          */
-        #getTokens() {
-            const allowedTypes = ['character', 'npc']
-            const tokens = coreModule.api.Utils.getControlledTokens()
-            const actors = tokens?.filter(token => token.actor).map((token) => token.actor)
-            if (actors.every((actor) => allowedTypes.includes(actor.type))) {
-                return tokens
-            } else {
-                return []
-            }
+        #getValidActors() {
+            const allowedTypes = ["character", "npc"];
+            return this.actors.every(actor => allowedTypes.includes(actor.type)) ? this.actors : [];
         }
+
+        /**
+         * Get valid tokens
+         * @private
+         * @returns {object}
+         */
+        #getValidTokens() {
+            const allowedTypes = ["character", "npc"];
+            return this.actors.every(actor => allowedTypes.includes(actor.type)) ? this.tokens : [];
+        }
+
+        /**
+        * Get action
+        * @private
+        * @param {object} entity      The entity
+        *  @param {string} actionType The action type
+        * @returns {object}           The action
+        */
+        async #getAction(entity, actionType = "item") {
+            var id = "", info = {}, tooltip = "";
+            switch (actionType) {
+                case "skill":
+                    id = entity.value;
+                    break;
+                case "weapon":
+                case "shipweapon":
+                    id = entity._id;
+                    //info = this.#getItemInfo(entity);
+                    tooltip = await this.#getTooltip(await this.#getWeaponTooltipData(entity));
+                    break;
+            }
+
+            let name = entity?.name ?? entity?.label;
+            let cssClass = "";
+
+            return {
+                id,
+                name,
+                cssClass,
+                img: coreModule.api.Utils.getImage(entity),
+                // icon1: this.#getActivationTypeIcon(entity.system?.activities?.contents[0]?.activation.type),
+                // icon2: this.#getPreparedIcon(entity),
+                // icon3: this.#getConcentrationIcon(entity),
+                // info1: info?.info1,
+                // info2: info?.info2,
+                // info3: info?.info3,
+                listName: this.#getListName(actionType, name),
+                tooltip,
+                system: { actionType, actionId: id }
+            };
+        }
+
+        /**
+     * Is equipped item
+     * @private
+     * @param {object} item The item
+     * @returns {boolean}   Whether the item is equipped
+     */
+        #isEquippedItem(item) {
+            const excludedTypes = ["gear"];
+            return (this.displayUnequipped && !excludedTypes.includes(item.type))
+                || (item.system?.equippable?.equipped && item.type !== "consumable");
+        }
+        /**
+         * Build actions
+         * @public
+         * @param {object} data actionData, groupData, actionType
+         * @param {object} options
+         */
+        async buildActions(data, options) {
+            const { actionData, groupData, actionType } = data;
+            // Exit if there is no action data
+            if (actionData.size === 0) return;
+
+            // Exit if there is no groupId
+            const groupId = (typeof groupData === "string" ? groupData : groupData?.id);
+            if (!groupId) return;
+
+            // Get actions
+            const actions = await Promise.all([...actionData].map(async item => await this.#getAction(item[1], actionType)));
+
+            // Add actions to action list
+            this.addActions(actions, groupData);
+        }
+
+        #getListName(actionType, actionName) {
+            const prefix = `${game.i18n.localize(ACTION_TYPE[actionType])}: ` ?? "";
+            return `${prefix}${actionName}` ?? "";
+        }
+
+        async #getWeaponTooltipData(entity) {
+            if (this.tooltipsSetting === "none") return "";
+
+            const name = entity?.name ?? "";
+
+            if (this.tooltipsSetting === "nameOnly") return name;
+            const damage = entity?.system?.damage.adjusted
+            const crit = entity?.system?.crit.adjusted
+            const range = entity?.system?.range.label;
+            //const skill = entity?.system?.skill.value;
+            //const description = (typeof entity?.system?.description === "string") ? entity?.system?.description : (unidentified ? entity?.system?.unidentified?.description : entity?.system?.description?.value) ?? "";
+
+            return { name, damage, crit, range };
+        }
+
+        /**
+         * Get tooltip
+         * @param {object} tooltipData The tooltip data
+         * @returns {string}           The tooltip
+         */
+        async #getTooltip(tooltipData) {
+            if (this.tooltipsSetting === "none") return "";
+            if (typeof tooltipData === "string") return tooltipData;
+
+            const name = game.i18n.localize(tooltipData.name);
+
+            if (this.tooltipsSetting === "nameOnly") return name;
+
+            const nameHtml = `<h3>${name}</h3>`;
+
+            // const skillHtml = tooltipData?.skill
+            //     ? `<span class="tah-tag}">${game.i18n.localize("SWFFG.ItemWeaponSkill")} ${game.i18n.localize(tooltipData.skill)}</span>`
+            //     : "";
+
+            const rangeHtml = tooltipData?.range
+                ? `<div class="tah-properties">${game.i18n.localize("SWFFG.ItemsRange")} ${game.i18n.localize(tooltipData.range)}</div>`
+                : "";
+
+            const damageHtml = tooltipData?.damage
+                ? `<div class="tah-properties">${game.i18n.localize("SWFFG.ItemsDamage")} ${tooltipData?.damage}</div>`
+                : "";
+
+            const critHtml = tooltipData?.damage
+                ? `<div class="tah-properties">${game.i18n.localize("SWFFG.ItemsCrit")} ${tooltipData?.crit}</div>`
+                : "";
+
+
+            const tagsJoined = [rangeHtml, damageHtml, critHtml].join("");
+
+            const tagsHtml = (tagsJoined) ? `<div class="tah-tags">${tagsJoined}</div>` : "";
+
+            const headerTags = (tagsHtml) ? `<div class="tah-tags-wrapper">${tagsHtml}</div>` : "";
+
+            if (!tagsHtml) return name;
+
+            return `<div>${nameHtml}${headerTags}</div>`;
+        }
+
+        /**
+         * Get item info
+         * @private
+         * @param {object} item
+         * @returns {object}
+         */
+        #getItemInfo(item) {
+            const damage = item?.system?.damage.adjusted
+            const crit = item?.system?.crit.adjusted
+            const info1 = { text: game.i18n.localize(item?.system?.range.label), title: "SWFFG.ItemsRange" };
+            const info2 = { damage, title: "SWFFG.ItemsDamage" };
+            const info3 = { crit, title: "SWFFG.ItemsCrit" };
+
+            return { info1, info2, info3 };
+        }
+
+
     }
+
+
+
+
 })