+++ scripts/settings.js
@@ -5,7 +5,7 @@ import { MODULE } from './constants.js'
  * Called by Token Action HUD Core to register Token Action HUD system module settings
  * @param {function} coreUpdate Token Action HUD Core update function
  */
-export function register (coreUpdate) {
+export function register(coreUpdate) {
     game.settings.register(MODULE.ID, 'displayUnequipped', {
         name: game.i18n.localize('tokenActionHud.template.settings.displayUnequipped.name'),
         hint: game.i18n.localize('tokenActionHud.template.settings.displayUnequipped.hint'