+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+`12.2.1.1_903`
+
+* Features:
+  * Add tooltip for weapons (Range, damage and critical values)
+  * Use the new dialog system (light or dark theme)
+  * For crew skills using the Pilot role or handling, difficulty and challenge dice are automatically placed if the current speed is greater than 0
+  * If the vehicle has more than 1 weapon open a dialog to choose the weapon after use a crew action with a use weapon role
+* Fixes:
+  * Refactoring code and remove deprecated functions for TAH version 2.x
+  * Skill name is translated for the Dice pool dialog and the chat message
+
 `12.2.0.1_903`
 
 * Features: